+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Add Input: MH-Z14 CO2
  - Add ability to set Chirp Input I2C address ([#1197](https://github.com/kizniche/mycodo/issues/1197))
  - Add UI feedback for Atlas Scientific calibration commands ([#1198](https://github.com/kizniche/Mycodo/issues/1198))
+ - Add ability to use non-standard condition/action ID strings in Conditional Functions
 
 ### Miscellaneous